Starting to make a superclass for the generators.
