+++ PanoramAI/__init__.py
@@ -1,6 +1,6 @@
 from .GANorama import GANorama
 from .VAEorama import VAEorama
-from .DataLoader import load_data
+from .generic import GENERICorama
 
 __author__ = "Tom McClintock <thmsmcclintock@gmail.com"
-__version__ = "0.1.0"
+__version__ = "0.1.1"
